+++ changelog.md
@@ -14,6 +14,7 @@
       - Add functions to rooms
       - Add standard object support to add money, gems, and standard objects
       - Add argument support to objects and monsters to allow passing arguments when object is created
+      - Hide water properties unless waterdrink flag enabled
     - Virtual area editor: Add ctrl+click selection system
 - **Fixed:**
   - Increase progress dialog height